# packages/api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: Mail Tracker API
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - owner: []
paths:
...
  /addon/pair:
    post:
      tags: [tracking]
      operationId: pairAddon
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddonPairInput' }
      responses:
        '200':
          description: paired
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AddonPairResult' }
  /addon/prepare:
    post:
      tags: [tracking]
      operationId: prepareAddonDraft
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddonPrepareInput' }
      responses:
        '200':
          description: pixel URL to insert
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AddonPrepareResult' }
...
  /mailboxes/connect-read:
    {
      'post':
        {
          'operationId': 'connectMailboxRead',
          'responses':
            {
              '200':
                {
                  'description': 'Successful operation',
                  'content':
                    {
                      'application/json':
                        { 'schema': { '$ref': '#/components/schemas/ConnectURL' } },
                    },
                },
              '400': { 'description': 'Invalid input' },
              '401': { 'description': 'Authentication required' },
              '404': { 'description': 'Resource not found' },
              '429': { 'description': 'Rate limit reached' },
              '500': { 'description': 'Internal error' },
            },
          'requestBody':
            {
              'required': true,
              'content':
                {
                  'application/json':
                    {
                      'schema':
                        {
                          'type': 'object',
                          'required': ['mailboxId'],
                          'properties': { 'mailboxId': { 'type': 'string' } },
                        },
                    },
                },
            },
        },
    }
...
  /mailboxes/{mailboxId}/pause-sync:
    {
      'post':
        {
          'operationId': 'pauseMailboxSync',
          'responses':
            {
              '200':
                {
                  'description': 'Successful operation',
                  'content':
                    {
                      'application/json':
                        {
                          'schema':
                            {
                              'type': 'object',
                              'required': ['paused'],
                              'properties': { 'paused': { 'type': 'boolean' } },
                            },
                        },
                    },
                },
              '400': { 'description': 'Invalid input' },
              '401': { 'description': 'Authentication required' },
              '404': { 'description': 'Resource not found' },
              '429': { 'description': 'Rate limit reached' },
              '500': { 'description': 'Internal error' },
            },
          'parameters':
            [
              {
                'name': 'mailboxId',
                'in': 'path',
                'required': true,
                'schema': { 'type': 'string' },
              },
            ],
        },
    }
  /conversations/{conversationId}/messages:
    {
      'get':
        {
          'operationId': 'listConversationMessages',
          'responses':
            {
              '200':
                {
                  'description': 'Successful operation',
                  'content':
                    {
                      'application/json':
                        {
                          'schema':
                            {
                              'type': 'array',
                              'items': { '$ref': '#/components/schemas/SyncedMessage' },
                            },
                        },
                    },
                },
              '400': { 'description': 'Invalid input' },
              '401': { 'description': 'Authentication required' },
              '404': { 'description': 'Resource not found' },
              '429': { 'description': 'Rate limit reached' },
              '500': { 'description': 'Internal error' },
            },
          'parameters':
            [
              {
                'name': 'conversationId',
                'in': 'path',
                'required': true,
                'schema': { 'type': 'string' },
              },
            ],
        },
    }
...
components:
  securitySchemes:
    owner: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
...
    ConnectURL:
      type: object
      required: [url]
      properties:
        url: { type: string }
...
    AddonPairInput:
      type: object
      required: [code, identityToken]
      properties:
        code: { type: string }
        identityToken: { type: string }
    AddonPairResult:
      type: object
      required: [mailboxId]
      properties:
        mailboxId: { type: string }
    AddonPrepareInput:
      type: object
      required: [identityToken, subject, recipients]
      properties:
        identityToken: { type: string }
        subject: { type: string }
        recipients: { type: array, items: { type: string } }
    AddonPrepareResult:
      type: object
      required: [conversationId, pixelUrl]
      properties:
        conversationId: { type: string }
        pixelUrl: { type: string }
...
    SyncedMessage:
      {
        'type': 'object',
        'required': ['id', 'subject', 'sender', 'snippet', 'direction', 'at'],
        'properties':
          {
            'id': { 'type': 'string' },
            'subject': { 'type': 'string' },
            'sender': { 'type': 'string' },
            'snippet': { 'type': 'string' },
            'direction': { 'type': 'string' },
            'at': { 'type': 'string', 'format': 'date-time' },
          },
      }